+++ sample/src/main/java/testlibuiza/sample/v3/customskin/CustomSkinXMLActivity.java
@@ -149,9 +149,9 @@ public class CustomSkinXMLActivity extends BaseActivity implements UZCallback {
 
     @Override
     public void isInitResult(boolean isInitSuccess, boolean isGetDataSuccess, ResultGetLinkPlay resultGetLinkPlay, Data data) {
-        if (isInitSuccess) {
+        /*if (isInitSuccess) {
             uzVideo.setEventBusMsgFromActivityIsInitSuccess();
-        }
+        }*/
     }
 
     @Override