+++ sample/src/main/java/testlibuiza/sample/v3/demoui/FrmVideoTop.java
@@ -288,7 +288,7 @@ public class FrmVideoTop extends BaseFragment implements UZCallback {
         ((HomeV4CanSlideActivity) getActivity()).isInitResult(isGetDataSuccess, resultGetLinkPlay, data);
         if (isInitSuccess) {
             setListener();
-            uzVideo.setEventBusMsgFromActivityIsInitSuccess();
+            //uzVideo.setEventBusMsgFromActivityIsInitSuccess();
         }
     }