+++ sample/src/main/java/testlibuiza/sample/v3/linkplay/PlayerActivity.java
@@ -1,0 +1,127 @@
+package testlibuiza.sample.v3.linkplay;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+
+import testlibuiza.R;
+import testlibuiza.app.LSApplication;
+import uizacoresdk.util.UZUtil;
+import uizacoresdk.view.rl.video.UZCallback;
+import uizacoresdk.view.rl.video.UZVideo;
+import vn.uiza.core.base.BaseActivity;
+import vn.uiza.core.common.Constants;
+import vn.uiza.restapi.uiza.model.v2.listallentity.Item;
+import vn.uiza.restapi.uiza.model.v3.linkplay.getlinkplay.ResultGetLinkPlay;
+import vn.uiza.restapi.uiza.model.v3.metadata.getdetailofmetadata.Data;
+
+/**
+ * Created by loitp on 7/16/2018.
+ */
+
+public class PlayerActivity extends BaseActivity implements UZCallback {
+    private UZVideo uzVideo;
+
+    @Override
+    protected boolean setFullScreen() {
+        return false;
+    }
+
+    @Override
+    protected String setTag() {
+        return "TAG" + getClass().getSimpleName();
+    }
+
+    @Override
+    protected int setLayoutResourceId() {
+        return R.layout.player_activity;
+    }
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        UZUtil.setCasty(this);
+        super.onCreate(savedInstanceState);
+        uzVideo = (UZVideo) findViewById(R.id.uiza_video);
+        uzVideo.setUZCallback(this);
+        UZUtil.initEntity(activity, uzVideo, LSApplication.entityIdDefaultVOD);
+    }
+
+    @Override
+    public void isInitResult(boolean isInitSuccess, boolean isGetDataSuccess, ResultGetLinkPlay resultGetLinkPlay, Data data) {
+        /*if (isInitSuccess) {
+            uzVideo.setEventBusMsgFromActivityIsInitSuccess();
+        }*/
+    }
+
+    @Override
+    public void onClickListEntityRelation(Item item, int position) {
+
+    }
+
+    @Override
+    public void onClickBack() {
+
+    }
+
+    @Override
+    public void onClickPip(Intent intent) {
+
+    }
+
+    @Override
+    public void onClickPipVideoInitSuccess(boolean isInitSuccess) {
+
+    }
+
+    @Override
+    public void onSkinChange() {
+
+    }
+
+    @Override
+    public void onError(Exception e) {
+
+    }
+
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+        uzVideo.onDestroy();
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        uzVideo.onResume();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        uzVideo.onPause();
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        uzVideo.onStart();
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        uzVideo.onStop();
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == Constants.CODE_DRAW_OVER_OTHER_APP_PERMISSION) {
+            if (resultCode == Activity.RESULT_OK) {
+                uzVideo.initializePiP();
+            }
+        } else {
+            super.onActivityResult(requestCode, resultCode, data);
+        }
+    }
+}